add appCoordinator
